requirements.txt
fastapi
fitz 
openai
fpdf
uvicorn
python-multipart
jinja2
python-docx